backend/pyproject.toml
[tool.poetry]
name = "clutter"
version = "0.1.0"
description = "Clutter image sharing backend"
authors = ["Matthew McAllister <matthew.mcallister.0@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.1"
jsonschema = "^4.21.1"
werkzeug = "^3.0.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
